Rendering the Collection... Better
But let's clean this up - because the form looks awful... even by my standards.

Open the template: app/Resources/views/admin/genus/_form.html.twig.

genusScientists field is not and actual field anymore: it's an array of fields. In fact, each of those field is itself composed of more sub-fields. What we have is a fairly complex form tree, which is something we talked about in our Form Theming Tutorial(https://knpuniversity.com/screencast/symfony-form-theming/compound-embedded-forms).

To render this in a more controlled way, delete the form_row. Then, add an h3 called 'Scientists', a Bootstrap row, and then loop over the fields with for genusScientistForm in genusForm.genusScientists.

Yep, we're looping over each of those four embedded forms.

Add a column, and then call form_row(genusScientistForm) to print both the user and yearsStudied fields at once.
So this should render the same thing as before, but with a bit more styling. Refresh! Ok, it's better... but what's up with those zero, one, two, three labels?

This genusScientistForm is actually an entire form full of several fields. So, it prints out a label for the entire form... which is zero, one, two, three, and four. That's not helpful!