Right now, this Genus is related to four GenusScientists. Cool... but what if one of those users stopped studying the Genus - how could we remove that one? The Delete UI & JavaScript
Let's plan out the UI first: I want to be able click a little x icon next to each embedded form to make it disappear from the page. Then, when we submit, it should fully delete that GenusScientist record from the database. Cool?

Inside the embedded form, add a new class to the column: js-genus-scientist-item.
We'll use that in JavaScript in a second. Below that, add a little link with its own js-remove-scientist class... and put the cute little 'x' icon inside