Oh, but back in _form.html.twig, add one more class to the row that's around the entire section called js-genus-scientist-wrapper. Ok, back to the JavaScript! Add var {dollar_sign}wrapper = then use jQuery to select that wrapper element. Register a listener on click for any .js-remove-scientist element - that's the delete link. Start that function with my favorite e.preventDefault(). Then... what next? Well, forget about Symfony and the database: just find the .js-genus-scientist-item element that's around this link and... remove it! Simple! Refresh the page, click that 'x', and be amazed.
Missing Fields: The allow_delete Option

But this is superficial: it didn't delete anything from the database nor can we submit the form and expect something to magically delete this GenusScientist, just because we removed it from the page. Or can we?

Submit! Well, I guess not. Huge error from the database!

    UPDATE genus_scientist SET years_studied and user_id to null.

Hmm. So our form is not expecting this embedded form to simply disappear. Instead, because the fields are missing from the submitted data, it thinks that we want to set that Genus Scientist's yearsStudied and user fields to null! No! I want to delete that entire object from the database!