Our form is not expecting this embedded form to simply disappear. Instead, because the fields are missing from the submitted data, it thinks that we want to set that Genus Scientist's yearsStudied and user fields to null! No! I want to delete that entire object from the database!
How can we do that? First, in GenusFormType, we need to tell the genusScientists field that it's ok if one of the embedded form's fields is missing from the submit. Set a new allow_delete option to true.

This tells the CollectionType that it's ok if one of the GenusScientist forms is missing when we submit. And, if a GenusScientist form is missing, it should remove that GenusScientist from the genusScientists array property. In other words, when we remove a GenusScientist form and submit, the final array will have three GenusScientist objects in it, instead of four.

Ready? Submit!

Hmm, no error... but it still doesn't work. Why not? Hint: we already know the answer... and it relates to Doctrine's inverse relationships. Let's fix it.