When we delete one of the GenusScientist forms and submit, the CollectionType is now smart enough to remove that GenusScientist from the genusScientists array on Genus. So, why doesn't that make any difference to the database?
The problem is that the genusScientists property is now the inverse side of this relationship.

In other words, if we remove or add a GenusScientist from this array, it doesn't make any difference! Doctrine ignores changes to the inverse side.
Setting the Owning Side: by_reference

How to fix it? We already know how! We did it back with our ManyToMany relationship! It's a two step process.

First, in GenusFormType, set the by_reference option to false.
Without this, the form component never calls setGenusScientists(). In fact, there is no setGenusScientists method in Genus. Instead, the form calls getGenusScientists() and then modifies that ArrayCollection object by reference.
But by setting it to false, it's going to give us the flexibility we need to set the owning side of the relationship.