Setting the Owning Side: Adder & Remover
With just set the by_reference option to false change, submit the form. Error! But look at it closely: the error happens when the form system calls removeGenusScientist(). That's perfect! Well, not the error, but when we set by_reference to false, the form started using our adder and remover methods. Now, when we delete a GenusScientist form, it calls removeGenusScientist().

The only problem is that those methods are totally outdated: they're still written for our old ManyToMany setup.

In removeGenusScientist(), change the argument to accept a GenusScientist object. Then update {dollar_sign}user to {dollar_sign}genusScientist in one spot, and then the other.

For the last line, use {dollar_sign}genusScientist->setGenus(null). Let's update the note to say the opposite:

    Needed to update the owning side of the relationship!

Now, when we remove one of the embedded GenusScientist forms and submit, it will call removeGenusScientist() and that will set the owning side: {dollar_sign}genusScientist->setGenus(null).

If you're a bit confused how this will ultimately delete that GenusScientist, hold on! Because you're right! But, submit the form again.

The Missing Link: orphanRemoval

Yay! Another error!

    UPDATE genus_scientist SET genus_id = NULL

Huh... that makes perfect sense. Our code is not deleting that GenusEntity. Nope, it's simply setting its genus property to null. This update query makes sense!

But... it's not what we want! We want to say:

    No no no. If the GenusScientist is no longer set to this Genus, it should be deleted entirely from the database.