And Doctrine has an option for exactly that. In Genus, find your genusScientists property. Let's reorganize the OneToMany annotation onto multiple lines: it's getting a bit long. Then, add one magical option: orphanRemoval = true.
That's the key. It says:

    If one of these GenusScientist objects suddenly has their genus set to null, just delete it entirely.

Tip

If the GenusScientist.genus property is set to a different Genus, instead of null, it will still be deleted. Use orphanRemoval only when that's not going to happen.

Give it a try! Refresh the form to start over. We have four genus scientists. Remove one and hit save.

Woohoo! That fourth GenusScientist was just deleted from the database.

I know this was a bit tricky, but we didn't write a lot of code to get here. There are just two things to remember.

First, if you're ever modifying the inverse side of a relationship in a form, set by_reference to false, create adder and remover methods, and set the owning side in each. And second, for a OneToMany relationship like this, use orphanRemoval to delete that related entity for you.

This was a big success! Next: we need to be able to add new genus scientists in the form.