So, how can we add a new scientist to a Genus?
Here's the plan: I want to add a button called 'Add New Scientist', and when the user clicks it, it will render a new blank, embedded GenusScientist form. After the user fills in those fields and saves, we will insert a new record into the genus_scientist table.

The allow_add Option

Let's start with the front end first. Open GenusFormType. After the allow_delete option, put a new one: allow_add set to true.

Remember: allow_delete says:

It's ok if one of the genus scientists' fields are missing from the submitted data.
And when one is missing, the form should remove it from the genusScientists array.

The allow_add option does the opposite:

If there is suddenly an extra set of GenusScientist form data that's submitted, that's great!
In this case, it will create a new GenusScientist object and set it on the genusScientists array.