Using... the prototype!
No worries! Symfony's CollectionType has a crazy thing to help us: the prototype.

Google for 'Symfony form collection' and open the How to Embed a Collection of Forms(https://symfony.com/doc/current/form/form_collections.html) document on Symfony.com. This page has some code that's ripe for stealing!

First, under the 'Allowing New' section, find the template and copy the data-prototype attribute code. Open our form template, and add this to the wrapper div. Update the variable to genusForm.genusScientists.vars.prototype.
Add one other thing while we're here: I promise I'll explain all of this in a minute: data-index set to genusForm.genusScientists|length.

Checking out the prototype: __name__

View the HTML source and search for wrapper to find our js-genus-scientist-wrapper element. That big mess of characters is the prototype. Yep, it looks crazy. This is a blank version of one of these embedded forms... after being escaped with HTML entities so that it can safely live in an attribute. This is great, because we can read this in JavaScript when the user clicks 'Add New Scientist'.

Oh, but check out this __name__ string: it shows up in a bunch of places inside the prototype. Scroll down a little to the embedded GenusScientist forms. If you look closely, you'll see that the fields in each of these forms have a different index number. The first is index zero, and it appears in a few places, like the name and id attributes. The next set of fields use one and then two.

When Symfony renders the prototype, instead of hard coding a number there - like zero, one or two - it uses __name__. It then expects us - in JavaScript - to change that to a unique index number, like three.

The Prototype JavaScript

Let's do it! Back on the Symfony documentation page: a lot of the JavaScript we need lives here. Find the addTagForm() function and copy the inside of it. Back in edit.html.twig, paste this inside our click function.

And let's make some changes. First, update {dollar_sign}collectionHolder to {dollar_sign}wrapper: that's the element that has the data-prototype attribute. We also read the data-index attribute... which is important because it tells us what number to use for the index. This is used to replace __name__ with that number. And then, each time we add another form, this index goes up by one.

Finally, at the very bottom: put this new sub-form onto the page: {dollar_sign}(this) - which is the 'Add another Scientist' link, {dollar_sign}(this).before(newForm).

I think we are ready! Find your browser and refresh! Hold your breath: click 'Add Another Scientist'. It works! Well, the styling isn't quite right... but hey, this is a victory! And yea, we'll fix the styling later.

Add one new scientist, and hit save. Ah! It blows up! Obviously, we have a little bit more work to do.