We just need to call persist() on any new GenusScientist objects. We could add some code to our controller to do that after the form is submitted: class GenusAdminController extends Controller { ... lines 18 - 34     public function newAction(Request )     {         {dollar_sign}form = {dollar_sign}this->createForm(GenusFormType::class);         // only handles data on POST         {dollar_sign}form->handleRequest({dollar_sign}request);         if ({dollar_sign}form->isSubmitted() && {dollar_sign}form->isValid()) { ... lines 42 - 43             {dollar_sign}em = ->getDoctrine()->getManager();             {dollar_sign}em->persist({dollar_sign}genus);             {dollar_sign}em->flush(); ... lines 47 - 53         } ... lines 55 - 58     } ... lines 60 - 87 }
Or... we could do something fancier. In Genus, add a new option to the OneToMany: cascade={'persist'}

This says:

When we persist a Genus, automatically call persist on each of the GenusScientist objects in this array. In other words, cascade the persist onto these children.
Alright, refresh now. This is the last error, I promise! And this makes perfect sense: it is trying to insert into genus_scientist - yay! But with genus_id set to null.