We've got more work to do! So head back to /admin/genus. Leave the 'Years Studied' field empty for one of the GenusScientist forms and submit.
Explosion!

UPDATE genus_scientist SET years_studied = NULL
This field is not allowed to be null in the database. That's on purpose... but we're missing validation!

But no problem, right? We'll just go into the Genus class, copy the as Assert use statement, paste it into GenusScientist and then - above yearsStudied - add @Assert\NotBlank.

@Valid for a Good Time

It's as if Symfony doesn't see the new validation constraint! Why? Here's the deal: our form is bound to a Genus object.

That's the top-level object that we're modifying. And by default, Symfony reads all of the validation annotations from the top-level class... only. When it sees an embedded object, or an array of embedded objects, like the genusScientists property, it does not go deeper and read the annotations from the GenusScientist class. In other words, Symfony only validates the top-level object.

Double-lame! What the heck Symfony?

No no, it's cool, it's on purpose. You can easily activate embedded validation by adding a unique annotation above that property: @Assert\Valid
That's it! Now refresh. Validation achieved!