There's still one ugly problem with our form, and I promised we would fix: when we click 'Add Another Scientist'... well, it don't look right... lines 24 - 28 The new form should have the exact same styling as the existing ones.
Customizing the Prototype!

Why does it look different, anyways? Remember the data-prototype attribute?
By calling form_widget, this renders a blank GenusScientist form... by using the default Symfony styling. But when we render the existing embedded forms, we wrap them in all kinds of cool markup.

What we really want is to somehow make the data-prototype attribute use the markup that we wrote inside the for statement.

How? Well, there are at least two ways of doing it, and I'm going to show you the less-official and - in my opinion - easier way!

Head to the top of the file and add a macro called printGenusScientistRow() that accepts a genusScientistForm argument.
If you haven't seen a macro before in Twig, it's basically a function that you create right inside Twig. It's really handy when you have some markup that you don't want to repeat over and over again.