POST_SET_DATA is a constant for an event called form.post_set_data. This is called after the data behind the form is added to it: in other words, after the GenusScientist is bound to each embedded form.
When that happens, the form system will call an onPostSetData() function, which we create: public function onPostSetData(). This will receive a FormEvent object.
Now we're close! Inside, add an if statement: if {dollar_sign}event->getData():This form is always bound to a GenusScientist object. So this will return the GenusScientist object bound to this form, or - if this is a new form - then it may return null. That's why we'll say if {dollar_sign}event->getData() && {dollar_sign}event->getData()->getId().

In human-speak: as long as there is a GenusScientist bound to this form and it's been saved to the database - i.e. it has an id value - then let's unset the user field from the form.

To do that, fetch the form with {dollar_sign}form = {dollar_sign}event->getForm(). Then, literally, unset({dollar_sign}form['user'])

This {dollar_sign}form variable is a Form object, but you can treat it like an array, including unsetting fields.

That's it for the form! The last step is to conditionally render the user field. Because if we refresh right now, the form system yells at us:

There's no user field inside of our template at line 9.