The last step is to conditionally render the user field. Because if we refreshed before, the form system yelled at us:
There's no user field inside of our template at line 9.

Wrap that in an if statement: if genusScientistForm.user is defined, then print it.

lse, use a strong tag and print the user's e-mail address with genusScientistForm.vars - which is something we mastered in our Form Theming tutorial - .data - which will be a GenusScientist object - .user.email
This says: find the GenusScientist object behind this form, call getUser() on it, and then call getEmail() on that.

I think it's time to celebrate! Refresh the form. It looks exactly like I wanted. It's like my birthday! And when we add a new one, it still has the drop-down.