Here's my new wild idea: any scientist that has studied a genus for longer than 20 years should be considered an expert. So, in addition to the number of scientists I also want to print the number of expert scientists next to it.
Look inside the list template: we're printing this number by saying genus.genusScientists|length. In other words, call getGenusScientists().
Fetch the results, and then count them:

But how could we filter this to only return GenusScientist results that have studied the Genus for longer than 20 years?

It's easy! In Genus, create a new public function called getExpertScientists()

Then, we'll loop over all of the scientists to find the experts. And actually, we can do that very easily by saying {dollar_sign}this->getGenusScientists()->filter(), which is a method on the ArrayCollection object. Pass that an anonymous function with a GenusScientist argument. Inside, return {dollar_sign}genusScientist-getYearsStudied() > 20.
This will loop over all of the genus scientists and return a new ArrayCollection that only contains the ones that have studied for more than 20 years.
To print this in the template, let's add a new line, then {{ genus.expertScientists|length }} and then 'experts'

Try it now. Yes!

Click to check out the queries. It still makes a COUNT query for each row... but wait: it also queries for all of the genus_scientist results for each genus. That sucks! Even if a Genus only has two experts... we're fetching all of the data for all of its genus scientists.

Why? Well, as soon as we loop over genusScientists doctrine realizes that it needs to go and query for all of the genus scientists for this Genus. Then, we happily loop over them to see which ones have more than 20 yearsStudied.

This may or may not be a huge performance problem. If every Genus always has just a few scientists, no big deal! But if a Genus has hundreds of scientists, this page will grind to a halt while it queries for and hydrates all of those extra GenusScientist objects.

There's a better way: and it uses a feature in Doctrine that - until recently - even I didn't know existed. And I'm super happy it does.