Organizing Criteria into your Repository
But, to keep my code organized, I prefer to have all of my query logic inside of repository classes, including Criteria. No worries! Open GenusRepository and create a new static public function createExpertCriteria().

It would be better to put this method in GenusScientistRepository, since it operates on that entity.
Copy the criteria line from genus, paste it here and return it. Oh, and be sure you type the a on Criteria and hit tab so that PhpStorm autocompletes the use statement.

Back inside Genus, we can simplify things {dollar_sign}this->getGenusScientists()->matching(GenusRepository::createExpertCriteria())

Refresh that! Sweet! It works just like before.