Criteria in Query Builder
Another advantage of building the Criteria inside of your repository is that you can use it in a query builder. Imagine that we needed to query for all of the experts in the entire system. To do that we could create a new public function - findAllExperts().

Once again, this method should actually live in GenusScientistRepository, but the idea is exactly the same :).
But, I want to avoid duplicating the query logic that we already have in the Criteria!

No worries! Just return {dollar_sign}this->createQueryBuilder('genus') then, addCriteria(self::createExpertCriteria().
Finish with the normal getQuery() and execute().

Ok guys, that's it - that's everything. We just attacked the stuff that really frustrates people with Doctrine and Forms. Collections are hard, but if you understand the mapping and the inverse side reality, you write your code to update the mapping side from the inverse side, and understand a few things like orphanRemoval and cascade, everything falls into place.